in progress completed

+++ lib/service-context/service-context-api/src/main/java/bs/lib/service/context/api/repository/EventRepository.java
@@ -12,5 +12,5 @@ public interface EventRepository {
 
     List<SCEvent> selectAllByFilters(EventSearchFilter searchFilter);
 
-    Long countAllByFilters(EventSearchFilter searchFilter);
+    Integer countAllByFilters(EventSearchFilter searchFilter);
 }
